API

Contents

Index

Model Definitions

Combinators

Base.:+ — Method

Base.:+(m1::AbstractModel, m2::AbstractModel)

Combine two models to create a composite AddModel. This adds two models pointwise, i.e.

julia> m1 = Gaussian()
julia> m2 = Disk()
julia> visibility(m1+m2, 1.0, 1.0) == visibility(m1, 1.0, 1.0) + visibility(m2, 1.0, 1.0)
true

VLBISkyModels.added — Function

added(m1::AbstractModel, m2::AbstractModel)

Combine two models to create a composite AddModel. This adds two models pointwise, i.e.

julia> m1 = Gaussian()
julia> m2 = Disk()
julia> visibility(added(m1,m2), 1.0, 1.0) == visibility(m1, 1.0, 1.0) + visibility(m2, 1.0, 1.0)
true

VLBISkyModels.convolved — Function

convolved(m1::AbstractModel, m2::AbstractModel)

Convolve two models to create a composite ConvolvedModel.

julia> m1 = Ring()
julia> m2 = Disk()
julia> convolved(m1, m2)

VLBISkyModels.components — Function

components(m::AbstractModel)

Returns the model components for a composite model. This will return a Tuple with all the models you have constructed.

Example

julia> m = Gaussian() + Disk()
julia> components(m)
(Gaussian{Float64}(), Disk{Float64}())

VLBISkyModels.smoothed — Function

smoothed(m::AbstractModel, σ::Number)

Smooths a model m with a Gaussian kernel with standard deviation σ.

Notes

This uses convolved to created the model, i.e.

julia> m1 = Disk()
julia> m2 = Gaussian()
julia> convolved(m1, m2) == smoothed(m1, 1.0)
true

VLBISkyModels.CompositeModel — Type

abstract type CompositeModel{M1, M2} <: ComradeBase.AbstractModel

Abstract type that denotes a composite model. Where we have combined two models together.

Implementation

Any implementation of a composite type must define the following methods:

visibility_point

uv_combinator

imanalytic

visanalytic

ComradeBase.intensity_point if model intensity is IsAnalytic

intensitymap_numeric! if model intensity is NotAnalytic

intensitymap_numeric if model intensity is NotAnalytic

flux

radialextent

In addition there are additional optional methods a person can define if needed:

intensitymap_analytic! if model intensity is IsAnalytic (optional)

intensitymap_analytic if model intensity is IsAnalytic (optional)

visibilities_analytic if visanalytic is IsAnalytic (optional)

visibilities_numeric if visanalytic is Not Analytic (optional)

visibilities_analytic! if visanalytic is IsAnalytic (optional)

visibilities_numeric! if visanalytic is Not Analytic (optional)

VLBISkyModels.AddModel — Type

struct AddModel{T1, T2} <: VLBISkyModels.CompositeModel{T1, T2}

Pointwise addition of two models in the image and visibility domain. An end user should instead call added or Base.+ when constructing a model

Example

julia> m1 = Disk() + Gaussian()
julia> m2 = added(Disk(), Gaussian()) + Ring()

VLBISkyModels.ConvolvedModel — Type

struct ConvolvedModel{M1, M2} <: VLBISkyModels.CompositeModel{M1, M2}

Pointwise addition of two models in the image and visibility domain. An end user should instead call convolved. Also see smoothed(m, σ) for a simplified function that convolves a model m with a Gaussian with standard deviation σ.

Geometric and Image Models

VLBISkyModels.GeometricModel — Type

abstract type GeometricModel{T} <: ComradeBase.AbstractModel

A type that defines it is a geometric model. These are usually primitive models, and are usually analytic in Fourier and the image domain. As a result a user only needs to implement the following methods

visibility_point

intensity_point

radialextent

Note that if the geometric model isn't analytic then the usual methods listed in ComradeBase.AbstractModel for non-analytic models need to be implemented.

VLBISkyModels.ConcordanceCrescent — Type

struct ConcordanceCrescent{T} <: VLBISkyModels.GeometricModel{T}

Creates the ConcordanceCrescent model, i.e. a flat-top crescent with a displacment and a slash and shadow depth. Note this creates a crescent with unit flux. If you want a different flux please use the renomed modifier.

Fields

router: Outer radius of the crescent

rinner: Inner radius of the crescent (i.e. inside this radius there is a hole)

shift: Displacment of the inner disk radius

slash: Strength of the linear slash. Note that s∈[0.0,1.0] to ensure positivity in the image.

Notes

Unlike the Gaussian and Disk models this does not create the unit version. In fact, this model could have been created using the Disk and primitives by using VLBISkyModels's model composition functionality.

VLBISkyModels.Crescent — Function

Creates a Kamruddin and Dexter crescent model. This works by composing two disk models together.

Arguments

router: The radius of the outer disk

rinner: The radius of the inner disk

shift: How much the inner disk radius is shifted (positive is to the right)

floor: The floor of the inner disk 0 means the inner intensity is zero and 1 means it is a large disk.

VLBISkyModels.Disk — Type

Disk{T}() where {T}

Tophat disk geometrical model, i.e. the intensity profile

\[ I(x,y) = \begin{cases} \pi^{-1} & x^2+y^2 < 1 \\ 0 & x^2+y^2 \geq 0 \end{cases}\]

i.e. a unit radius and unit flux disk.

By default if T isn't given, Disk defaults to Float64

VLBISkyModels.SlashedDisk — Type

SlashedDisk{T}(slash::T) where {T}

Tophat disk geometrical model, i.e. the intensity profile

\[ I(x,y) = \begin{cases} \pi^{-1} & x^2+y^2 < 1 \\ 0 & x^2+y^2 \geq 0 \end{cases}\]

i.e. a unit radius and unit flux disk.

By default if T isn't given, Disk defaults to Float64

VLBISkyModels.ExtendedRing — Type

struct ExtendedRing{T} <: VLBISkyModels.GeometricModel{T}

A symmetric extended ring whose radial profile follows an inverse gamma distributions.

The formula in the image domain is given by

I(r,θ) = βᵅrᵅ⁻²exp(-β/r)/2πΓ(α)

where α = shape and β = shape+1

Note

We mainly use this as an example of a non-analytic Fourier transform (although it has a complicated expression)

Fields

shape: shape of the radial distribution

Note that if T isn't specified at construction then it defaults to Float64.

VLBISkyModels.Gaussian — Type

struct Gaussian{T} <: VLBISkyModels.GeometricModel{T}

Gaussian with unit standard deviation and flux.

By default if T isn't given, Gaussian defaults to Float64

VLBISkyModels.MRing — Type

struct MRing{T, V<:Union{AbstractArray{T, 1}, Tuple{Vararg{T, N}} where {N, T}}} <: VLBISkyModels.GeometricModel{T}

m-ring geometric model. This is a infinitely thin unit flux delta ring whose angular structure is given by a Fourier expansion. That is,

I(r,θ) = (2π)⁻¹δ(r-1)∑ₙ(αₙcos(nθ) - βₙsin(nθ))

The N in the type defines the order of the Fourier expansion.

Fields

α: Real Fourier mode coefficients

β: Imaginary Fourier mode coefficients

VLBISkyModels.Ring — Type

struct Ring{T} <: VLBISkyModels.GeometricModel{T}

A infinitely thin ring model, whose expression in the image domain is I(r,θ) = δ(r - 1)/2π i.e. a unit radius and flux delta ring.

By default if T isn't given, Gaussian defaults to Float64

VLBISkyModels.ParabolicSegment — Type

struct ParabolicSegment{T} <: VLBISkyModels.GeometricModel{T}

A infinitely thin parabolic segment in the image domain. The segment is centered at zero, with roots ±1 and a yintercept of 1.

Note that if T isn't specified at construction then it defaults to Float64.

VLBISkyModels.ContinuousImage — Type

ContinuousImage{A<:IntensityMap, P} <: AbstractModel
ContinuousImage(img::Intensitymap, kernel)

The basic continuous image model for VLBISkyModels. This expects a IntensityMap style object as its imag as well as a image kernel or pulse that allows you to evaluate the image at any image and visibility location. The image model is

I(x,y) = ∑ᵢ Iᵢⱼ κ(x-xᵢ, y-yᵢ)

where Iᵢⱼ are the flux densities of the image img and κ is the intensity function for the kernel.

VLBISkyModels.ZeroModel — Type

struct ZeroModel{T} <: ComradeBase.AbstractModel

Defines a model that is empty that is it return zero for everything.

Notes

This returns 0 by using FillArrays so everything should be non-allocating

VLBISkyModels.MultiComponentModel — Type

MultiComponentModel(beam::AbstractModel, fluxes::AbstractVector, x::AbstractVector, y::AbstractVector)

Build a model with a base model type beam where fluxes, x, y corresond to the flux, and positions of the components. This can be used to easily construct clean like models.

VLBISkyModels.PolarizedModel — Type

struct PolarizedModel{TI, TQ, TU, TV} <: ComradeBase.AbstractPolarizedModel

Wrapped model for a polarized model. This uses the stokes representation of the image.

Fields

I: Stokes I model

Q: Stokes Q Model

U: Stokes U Model

V: Stokes V Model

Image Pulses

VLBISkyModels.Pulse — Type

Pulse

Pixel response function for a radio image model. This makes a discrete sampling continuous by picking a certain smoothing kernel for the image.

Notes

To see the implemented Pulses please use the subtypes function i.e. subtypes(Pulse)

VLBISkyModels.DeltaPulse — Type

struct DeltaPulse{T} <: VLBISkyModels.Pulse

A dirac comb pulse function. This means the image is just the dicrete Fourier transform

VLBISkyModels.BSplinePulse — Type

$(TYPEDEF)

Uses the basis spline (BSpline) kernel of order N. These are the kernel that come from recursively convolving the tophat kernel

\[ B_0(x) = \begin{cases} 1 & |x| < 1 \\ 0 & otherwise \end{cases}\]

N times.

Notes

BSpline kernels have a number of nice properties:

Simple frequency response $\sinc(u/2)^N$

preserve total intensity

For N>1 these kernels aren't actually interpolation kernels however, this doesn't matter for us.

Currently only the 0,1,3 order kernels are implemented.

VLBISkyModels.RaisedCosinePulse — Type

struct RaisedCosinePulse{T} <: VLBISkyModels.Pulse

RaisedCosinePulse()
RaisedCosinePulse(rolloff)

Raised cosine pulse function. This tends to be a very flat response, where the roll off controls the speed of decay. By default we set rolloff = 0.5.

VLBISkyModels.BicubicPulse — Type

struct BicubicPulse{T} <: VLBISkyModels.Pulse

BicubicPulse(b = 0.5)

The bicubic pulse for imaging. This pulse tends to have a flat spectrum but for most values of b can produce negative intensities in an image. This is the pulse used in Broderick et al. 2020.

VLBISkyModels.Butterworth — Type

Butterworth{N}()
Butterworth{N, T}()

Construct a model that corresponds to the Butterworth filter of order N. The type of the output is given by T and if not given defaults to Float64

Model Image (non analytic FFT)

VLBISkyModels.create_cache — Function

create_cache(alg::AbstractFourierTransform, img::AbstractIntensityMap)

Creates a Fourier transform cache for a img using algorithm alg. For non-analytic visibility models this can significantly speed up computations.

Examples

julia> u,v = rand(100), rand(100)
julia> cache = create_cache(DFTAlg(u, v), IntensityMap(randn(50,50), 10.0, 10.0))
julia> cache = create_cache(NFFTAlg(u, v), IntensityMap(randn(50,50), 10.0, 10.0))
julia> cache = create_cache(FFTAlg(), IntensityMap(randn(50,50), 10.0, 10.0))

VLBISkyModels.update_cache — Function

update_cache(cache, img)

Update the Fourier transform cache. This will reuse an FFT/NFFT plan saving some computational time.

Note

This is an intenal method than an end user shouldn't have to usually call.

VLBISkyModels.modelimage — Function

modelimage(model::AbstractModel, image::AbstractIntensityMap, alg=FFTAlg())

Construct a ModelImage from a model, image and the optionally specified visibility algorithm alg

Notes

For analytic models this is a no-op and returns the model. For non-analytic models this creates a ModelImage object which uses alg to compute the non-analytic Fourier transform.

modelimage(model, cache::AbstractCach))

Construct a ModelImage from the model and using a precompute Fourier transform cache. You can optionally specify th which will compute the internal image buffer using the`.

Example

julia> m = ExtendedRing(10.0)
julia> cache = create_cache(DFTAlg(), IntensityMap(zeros(128, 128), 50.0, 50.0)) # used threads to make the image
julia> mimg = modelimage(m, cache, true)

Notes

For analytic models this is a no-op and returns the model.

modelimage(m;
           fovx=2*radialextent(m),
           fovy=2*radialextent(m),
           nx=512,
           ny=512,
           alg=FFTAlg(),
           pulse=ComradeBase.DeltaPulse(),
            )

Construct a ModelImage where just the model m is specified.

If fovx or fovy aren't given modelimage will guess a reasonable field of view based on the radialextent function. nx and ny are the number of pixels in the x and y direction. The pulse is the pulse used for the image and alg

Notes

For analytic models this is a no-op and returns the model.

modelimage(img::ContinuousImage, alg=NFFTAlg())

Create a model image directly using an image, i.e. treating it as the model. You can optionally specify the Fourier transform algorithm using alg

modelimage(img::ContinuousImage, cache::AbstractCache)

Create a model image directly using an image, i.e. treating it as the model. Additionally reuse a previously compute image cache. This can be used when directly modeling an image of a fixed size and number of pixels.

VLBISkyModels.uviterator — Function

uviterator(nx, dx, ny dy)

Construct the u,v iterators for the Fourier transform of the image with pixel sizes dx, dy and number of pixels nx, ny

If you are extending Fourier transform stuff please use these functions to ensure that the centroid is being properly computed.

VLBISkyModels.fouriermap — Function

fouriermap(m, x)

Create a Fourier or visibility map of a model m where the image is specified in the image domain by the pixel locations x and y

VLBISkyModels.ModelImage — Type

struct ModelImage{M, I, C} <: VLBISkyModels.AbstractModelImage{M}

Container for non-analytic model that contains a image cache which will hold the image, a Fourier transform cache, which usually an instance of a <: FourierCache.

Note

This is an internal implementation detail that shouldn't usually be called directly. Instead the user should use the exported function modelimage, for example

using Comrade
m = ExtendedRing(20.0, 5.0)

# This creates an version where the image is dynamically specified according to the
# radial extent of the image
mimg = modelimage(m) # you can also optionally pass the number of pixels nx and ny

# Or you can create an IntensityMap
img = intensitymap(m, 100.0, 100.0, 512, 512)
mimg = modelimage(m, img)

# Or precompute a cache
cache = create_cache(FFTAlg(), img)
mimg = modelimage(m, cache)

VLBISkyModels.DFTAlg — Type

DFTAlg

Uses a discrete fourier transform. This is not very efficient for larger images. In those cases NFFTAlg or FFTAlg are more reasonable. For small images this is a reasonable choice especially since it's easy to define derivatives.

VLBISkyModels.FFTAlg — Type

struct FFTAlg <: VLBISkyModels.FourierTransform

Use an FFT to compute the approximate numerical visibilities of a model. For a DTFT see DFTAlg or for an NFFT NFFTAlg

Fields

padfac: The amount to pad the image by. Note we actually round up to the nearest factor of 2, but this will be improved in the future to use small primes

VLBISkyModels.FFTCache — Type

struct FFTCache{A<:FFTAlg, P, I, Pu, S} <: VLBISkyModels.AbstractCache

The cache used when the FFT algorithm is used to compute visibilties. This is an internal type and is not part of the public API

VLBISkyModels.NFFTAlg — Type

NFFTAlg

Uses a non-uniform FFT to compute the visibilities. You can optionally pass uv which are the uv positions you will compute the NFFT at. This can allow for the NFFT plan to be cached improving performance

Fields

padfac: Amount to pad the image

m: Kernel size parameters. This controls the accuracy of NFFT you do not usually need to change this

σ: Over sampling factor. This controls the accuracy of NFFT you do not usually need to change this.

window: Window function for the NFFT. You do not usually need to change this

precompute: NFFT interpolation algorithm.

blocking: Flag block partioning should be used to speed up computation

sortNodes: Flag if the node should be sorted in a lexicographic way

storeDeconvolutionIdx: Flag if the deconvolve indices should be stored, Currently required for GPU

fftflags: Flag passed to inner AbstractFFT. The fastest FFTW is FFTW.MEASURE but takes the longest to precompute

VLBISkyModels.NUFTCache — Type

struct NUFTCache{A, P, M, PI, I} <: VLBISkyModels.AbstractCache

Internal type used to store the cache for a non-uniform Fourier transform (NUFT).

The user should instead create this using the create_cache function.

VLBISkyModels.ObservedNUFT — Type

struct ObservedNUFT{A<:VLBISkyModels.NUFT, T} <: VLBISkyModels.NUFT

Container type for a non-uniform Fourier transform (NUFT). This stores the uv-positions that the model will be sampled at in the Fourier domain, allowing certain transformtion factors (e.g., NUFT matrix) to be cached.

This is an internal type, an end user should instead create this using NFFTAlg or DFTAlg.

Modifiers

VLBISkyModels.modify — Function

modify(m::AbstractModel, transforms...)

Modify a given model using the set of transforms. This is the most general function that allows you to apply a sequence of model transformation for example

modify(Gaussian(), Stretch(2.0, 1.0), Rotate(π/4), Shift(1.0, 2.0), Renorm(2.0))

will create a asymmetric Gaussian with position angle π/4 shifted to the position (1.0, 2.0) with a flux of 2 Jy. This is similar to Flux's chain.

modify(img::IntensityMap, transforms...)

This modifies the img by applying the transforms... returning a transformed IntensityMap

Note

Unlike when modify is applied to a <:AbstractModel this returns an already modified image.

VLBISkyModels.basemodel — Function

basemodel(model::ModifiedModel)

Returns the ModifiedModel with the last transformation stripped.

Example

julia> basemodel(stretched(Disk(), 1.0, 2.0)) == Disk()
true

VLBISkyModels.unmodified — Function

unmodified(model::ModifiedModel)

Returns the un-modified model

Example

julia> m = stretched(rotated(Gaussian(), π/4), 2.0, 1.0)
julia> umodified(m) == Gaussian()
true

VLBISkyModels.renormed — Function

renormed(model, f)

Renormalizes the model m to have total flux f*flux(m). This can also be done directly by calling Base.:* i.e.,

julia> renormed(m, f) == f*M
true

VLBISkyModels.rotated — Function

rotated(model, ξ)

Rotates the model by an amount ξ in radians in the clockwise direction.

VLBISkyModels.posangle — Function

posangle(model)

Returns the rotation angle of the rotated model

VLBISkyModels.shifted — Function

shifted(model, Δx, Δy)

Shifts the model m in the image domain by an amount Δx,Δy in the x and y directions respectively.

VLBISkyModels.stretched — Function

stretched(model, α, β)

Stretches the model m according to the formula Iₛ(x,y) = 1/(αβ) I(x/α, y/β), where were renormalize the intensity to preserve the models flux.

VLBISkyModels.ModifiedModel — Type

struct ModifiedModel{M<:ComradeBase.AbstractModel, T<:Tuple} <: ComradeBase.AbstractModel

Container type for models that have been transformed in some way. For a list of potential modifiers or transforms see subtypes(ModelModifiers).

Fields

model: base model

transform: model transforms

VLBISkyModels.ModelModifier — Type

abstract type ModelModifier{T}

General type for a model modifier. These transform any model using simple Fourier transform properties. To modify a model you can use the ModifiedModel constructor or the modify function.

julia> visanalytic(stretched(Disk(), 2.0, 2.0)) == visanalytic(Disk())
true

To implement a model transform you need to specify the following methods:

transform_uv

transform_image

scale_uv

scale_image

radialextent

See thee docstrings of those methods for guidance on implementation details.

Additionally these methods assume the modifiers are of the form

I(x,y) -> fᵢ(x,y)I(gᵢ(x,y)) V(u,v) -> fᵥ(u,v)V(gᵥ(u,v))

where g are the transformimage/uv functions and f are the scaleimage/uv function.

VLBISkyModels.Stretch — Type

Stretch(α, β)
Stretch(r)

Stretched the model in the x and y directions, i.e. the new intensity is Iₛ(x,y) = 1/(αβ) I(x/α, y/β), where were renormalize the intensity to preserve the models flux.

Example

julia> modify(Gaussian(), Stretch(2.0)) == stretched(Gaussian(), 2.0, 1.0)
true

If only a single argument is given it assumes the same stretch is applied in both direction.

julia> Stretch(2.0) == Stretch(2.0, 2.0)
true

VLBISkyModels.Shift — Type

struct Shift{T} <: VLBISkyModels.ModelModifier{T}

Shifts the model by Δx units in the x-direction and Δy units in the y-direction.

Example

julia> modify(Gaussian(), Shift(2.0, 1.0)) == shifted(Gaussian(), 2.0, 1.0)
true

VLBISkyModels.Rotate — Type

Rotate(ξ)

Type for the rotated model. This is more fine grained constrol of rotated model.

Example

julia> modify(Gaussian(), Rotate(2.0)) == rotated(Gaussian(), 2.0)
true

VLBISkyModels.Renormalize — Type

struct Renormalize{T} <: VLBISkyModels.ModelModifier{T}

Renormalizes the flux of the model to the new value scale*flux(model). We have also overloaded the Base.:* operator as syntactic sugar although I may get rid of this.

Example

julia> modify(Gaussian(), Renormalize(2.0)) == 2.0*Gaussian()
true

Model Evaluation

For more docstrings on how to evaluate models see ComradeBase.

VLBISkyModels.amplitude — Function

amplitude(model, p)

Computes the visibility amplitude of model m at the coordinate p. The coordinate p is expected to have the properties U, V, and sometimes Ti and Fr.

If you want to compute the amplitudes at a large number of positions consider using the amplitudes function.

VLBISkyModels.amplitudes — Function

amplitudes(m::AbstractModel, u::AbstractArray, v::AbstractArray)

Computes the visibility amplitudes of the model m at the coordinates p. The coordinates p are expected to have the properties U, V, and sometimes Ti and Fr.

VLBISkyModels.bispectra — Function

bispectra(m, p1, p2, p3)

Computes the closure phases of the model m at the triangles p1, p2, p3, where pi are coordinates.

VLBISkyModels.bispectrum — Function

bispectrum(model, p1, p2, p3)

Computes the complex bispectrum of model m at the uv-triangle p1 -> p2 -> p3

If you want to compute the bispectrum over a number of triangles consider using the bispectra function.

VLBISkyModels.closure_phase — Function

closure_phase(model, p1, p2, p3, p4)

Computes the closure phase of model m at the uv-triangle u1,v1 -> u2,v2 -> u3,v3

If you want to compute closure phases over a number of triangles consider using the closure_phases function.

VLBISkyModels.closure_phases — Function

closure_phases(m,
               p1::AbstractArray
               p2::AbstractArray
               p3::AbstractArray
               )

Computes the closure phases of the model m at the triangles p1, p2, p3, where pi are coordinates.

VLBISkyModels.logclosure_amplitude — Function

logclosure_amplitude(model, p1, p2, p3, p4)

Computes the log-closure amplitude of model m at the uv-quadrangle u1,v1 -> u2,v2 -> u3,v3 -> u4,v4 using the formula

\[C = \log\left|\frac{V(u1,v1)V(u2,v2)}{V(u3,v3)V(u4,v4)}\right|\]

If you want to compute log closure amplitudes over a number of triangles consider using the logclosure_amplitudes function.

VLBISkyModels.logclosure_amplitudes — Function

logclosure_amplitudes(m::AbstractModel,
                      p1,
                      p2,
                      p3,
                      p4
                     )

Computes the log closure amplitudes of the model m at the quadrangles p1, p2, p3, p4.

VLBISkyModels.visibility — Function

visibility(mimg, p)

Computes the complex visibility of model m at coordinates p. p corresponds to the coordinates of the model. These need to have the properties U, V and sometimes Ti for time and Fr for frequency.

Notes

If you want to compute the visibilities at a large number of positions consider using the visibilities.

Internal (Not Public API)

VLBISkyModels.scale_uv — Function

scale_image(model::AbstractModifier, u, u)

Returns a number on how to scale the image visibility at u v for an modified model

VLBISkyModels.scale_image — Function

scale_image(model::AbstractModifier, x, y)

Returns a number of how to to scale the image intensity at x y for an modified model

VLBISkyModels.transform_uv — Function

transform_uv(model::AbstractModifier, u, u)

Returns a transformed u and v according to the model modifier

VLBISkyModels.transform_image — Function

transform_image(model::AbstractModifier, x, y)

Returns a transformed x and y according to the model modifier

VLBISkyModels.ThreadedModel — Type

ThreadedModel

Experimental model wrapper than enables multi-threading when evaluating intensitymap